requirements.txt
transformers==4.41.2
trl
peft
accelerate==0.27.2
bitsandbytes==0.41.1
huggingface-hub==0.23.3
jsonlines==4.0.0
pandas
protobuf==3.20.1
sentencepiece==0.2.0
tensorboard
torch
vllm==v0.4.3
vllm-flash-attn
nltk
fraction
PyYAML
typing_extensions==4.7.1
openai==0.28
auto-gptq==0.6.0
